+++ CHANGES.rst
@@ -4,6 +4,9 @@ Changelog
 2.4.4 (unreleased)
 ------------------
 
+- fix z3cform support - respect referer
+  [tmog / sunew]
+
 - fix z3cform support - fix for vocabularies, lifted from
   p.dexterity addForm. 
   [tmog / sunew]